Moved Authenticator in App\Api folder.
